Add radius determination and edit prepareMD
radius determination
- determine radius by first find the centroid of the solute - determine
the distance from the outer most atom in solute then from the one in
solvent
Add the two distance together - this is a crude way thinking the solute
atom will probably travel around the first radius and then the second
radius is the solvent. Extra 3 Angstrom is also added.

prepareMD2 is now update with 6-31+(2d,p) and spaces, blank line at the
bottom

fincut2 - is updated with a fix to 6-311++G** spaces, blank line at the
bottom